+++ java-io-base/src/main/java/internal/io/EncodingFileFormatter.java
@@ -39,7 +39,7 @@ public final class EncodingFileFormatter<T> implements FileFormatter<T> {
 
     @Override
     public void formatStream(@NonNull T value, @NonNull OutputStream resource) throws IOException {
-        try (OutputStream encoding = encoder.applyWithIO(resource)) {
+        try (OutputStream encoding = encoder.applyWithIO(new UncloseableOutputStream(resource))) {
             formatter.formatStream(value, encoding);
         }
     }