+++ java-io-base/src/main/java/internal/io/FunctionalFileFormatter.java
@@ -10,11 +10,10 @@ import java.io.OutputStream;
 @lombok.RequiredArgsConstructor
 public final class FunctionalFileFormatter<T> implements FileFormatter<T> {
 
-    @NonNull
-    private final IOBiConsumer<? super T, ? super OutputStream> function;
+    private final @NonNull IOBiConsumer<? super T, ? super OutputStream> function;
 
     @Override
     public void formatStream(@NonNull T value, @NonNull OutputStream resource) throws IOException {
-        function.acceptWithIO(value, resource);
+        function.acceptWithIO(value, new UncloseableOutputStream(resource));
     }
 }