+++ java-io-xml/src/main/java/nbbrd/io/xml/Stax.java
@@ -16,6 +16,7 @@
  */
 package nbbrd.io.xml;
 
+import internal.io.UncloseableInputStream;
 import internal.io.text.LegacyFiles;
 import lombok.NonNull;
 import nbbrd.design.StaticFactoryMethod;
@@ -173,12 +174,12 @@ public class Stax {
 
         @Override
         public @NonNull T parseStream(@NonNull InputStream resource) throws IOException {
-            return doParseOrClose(o -> o.createXMLStreamReader(resource), NOTHING_TO_CLOSE);
+            return doParseOrClose(o -> o.createXMLStreamReader(new UncloseableInputStream(resource)), NOTHING_TO_CLOSE);
         }
 
         @Override
         public @NonNull T parseStream(@NonNull InputStream resource, @NonNull Charset encoding) throws IOException {
-            return doParseOrClose(o -> o.createXMLStreamReader(resource, encoding.name()), NOTHING_TO_CLOSE);
+            return doParseOrClose(o -> o.createXMLStreamReader(new UncloseableInputStream(resource), encoding.name()), NOTHING_TO_CLOSE);
         }
 
         @NonNull
@@ -274,12 +275,12 @@ public class Stax {
 
         @Override
         public @NonNull T parseStream(@NonNull InputStream resource) throws IOException {
-            return doParseOrClose(o -> o.createXMLEventReader(resource), NOTHING_TO_CLOSE);
+            return doParseOrClose(o -> o.createXMLEventReader(new UncloseableInputStream(resource)), NOTHING_TO_CLOSE);
         }
 
         @Override
         public @NonNull T parseStream(@NonNull InputStream resource, @NonNull Charset encoding) throws IOException {
-            return doParseOrClose(o -> o.createXMLEventReader(resource, encoding.name()), NOTHING_TO_CLOSE);
+            return doParseOrClose(o -> o.createXMLEventReader(new UncloseableInputStream(resource), encoding.name()), NOTHING_TO_CLOSE);
         }
 
         private T doParseOrClose(StaxFunction<XMLInputFactory, XMLEventReader> supplier, Closeable onClose) throws IOException {